mock http for all local tests and separate them from remote tests
