+++ test/cases/css-multiline.html
@@ -8,7 +8,7 @@
     <!-- ended with ">" -->
     <link href="assets/main.css"
           rel="stylesheet">
-    <link href="/assets/main.css"
+    <link href="assets/main.css"
           rel="stylesheet">
     <link href="https://raw.githubusercontent.com/jrit/html-resource-inline/master/test/cases/assets/simple.css"
           rel="stylesheet">
@@ -19,13 +19,13 @@
 
 	<!-- Favicons
 	================================================== -->
-	<link rel="shortcut icon" href="">	
-	
-	
+	<link rel="shortcut icon" href="">
+
+
 <link rel='stylesheet' id='bootstrap-css'  href='https://raw.githubusercontent.com/jrit/html-resource-inline/master/test/cases/assets/simple.css' type='text/css' media='all' />
-<link rel='stylesheet' id='foobar-css'  href='/assets/main.css' type='text/css' media='all' />
+<link rel='stylesheet' id='foobar-css'  href='assets/main.css' type='text/css' media='all' />
     <!--[if (lt IE 9) & (!IEMobile)]>
-      <link rel="stylesheet" href="/assets/ie.css">
+      <link rel="stylesheet" href="assets/ie.css">
     <![endif]-->
 
 </head>
