+++ test/cases/css-multiline_out.html
@@ -45,9 +45,9 @@ body{font-weight: bold;}
 
 	<!-- Favicons
 	================================================== -->
-	<link rel="shortcut icon" href="">	
-	
-	
+	<link rel="shortcut icon" href="">
+
+
 <style id='bootstrap-css'   type='text/css' media='all'>
 body{font-weight: bold;}
